Extract byte order agnostic BaseWritableMemoryImpl
